Refactor tealogger structure
